- Added support for creating v5 GL nodes to ZDBSP.
SVN r159 (trunk)

+++ doomdata.h
@@ -94,8 +94,8 @@ struct MapSegGL
 
 struct MapSegGLEx
 {
-	WORD	v1;
-	WORD	v2;
+	DWORD	v1;
+	DWORD	v2;
 	WORD	linedef;
 	WORD	side;
 	DWORD	partner;
